Integrate python based powerline into zshkit.
This patch will force zsh to use powerline for its prompt information if
powerline is installed for the current user. If not it will fall back to
prompt-wunjo which also has git status support.

This patch also introduces the powerline fonts as a submodule for
convenient installation on a new target system. (If that target system
is a local computer.)