Disable powerline.
Powerline is a bit of a pain to support so I am going to try to do
something else. Honestly this is a colorscheme issue more than anything
but I would rather use a simpler prompt for this than deal with the
issues I am going to have with it. It will probably come back when my
terminal emulators update.